Changed the sourcing behaviour of pgbasenv.sh. Clusters are not automatically removed from $PGBASENV_BASE/etc/pgclustertab if the detection does not find them anymore. This change addresses also issue #7.
This improvement prevents clusters from being removed during a restore or standby reinstate operation where for example $PGDATA is temporarily deleted/unavailable. Since the automatic detection of clusters is based on files in $PGDATA they must be kept during a login.

However, there is an new parameter --clean-pgclustertab for pgbasenv.sh which forces the tool to remove clusters from pgclustertab which do not
have a valid $PGDATA directory anymore.