Exclude more directories from automatic scan
On modern systems, the list of places where a database should not
be is even longer than it ever was: special purpose directories,
the operating system environment and other installation paths.
Also, there are more filesystems that should not be touched: autofs
(as we do not want to trigger random automounts here), ramfs (which
can pop up in various places) and nfs4 (nfs is already excluded).